+++ src/index.ts
@@ -35,11 +35,11 @@ export * from "./widgets/multi-date-field/multi-date-field";
 export * from "./widgets/month-date-field/month-date-field";
 export * from "./widgets/time-picker/time-picker";
 export * from "./widgets/time-field/time-field";
+export * from "./widgets/file-picker/file-picker";
 
 export * from "./utils/use-query-params/use-query-params";
 export * from "./utils/use-countdown/use-countdown";
 export * from "./utils/use-window-breakpoints/use-window-breakpoints";
 export * from "./utils/debounce/debounce";
 
-// export * from "../src-old/form/file-picker";
 // export * from "../src-old/form/file-field";